requirements/linux.txt
certifi==2020.12.5
chardet==4.0.0
clipboard==0.0.4
docutils==0.16
evdev==1.3.0
idna==2.10
Kivy==2.0.0
Kivy-Garden==0.1.4
Pillow==8.0.1
Pygments==2.7.3
pyinstaller==4.3
pynput==1.6.8
pyperclip==1.8.1
python-xlib==0.29
requests==2.25.1
six==1.15.0
urllib3==1.26.2
